backend/python-services/requirements.txt
flask==3.0.0
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
joblib==1.3.2
python-dotenv==1.0.0
elevenlabs==2.38.1
